Separate out user domain
