+++ domain/__init__.py
@@ -1,0 +1,1 @@
+"""Pure domain logic (no Dash)."""
